pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "docx>=0.2.4",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "plotly>=6.1.2",
    "python-docx>=1.2.0",
    "reportlab>=4.4.2",
    "streamlit>=1.46.0",
    "timedelta>=2020.12.3",
    "trafilatura>=2.0.0",
]
